refactor: :recycle: Rename modules
Following best practices to name the modules and ordering correctly the packages when importing them
