+++ model/github_user_test.go
@@ -1,4 +1,4 @@
-package models
+package model
 
 import "testing"
